requirements.txt
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.9
certifi==2025.1.31
dill==0.3.9
distro==1.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isort==6.0.1
jiter==0.9.0
mccabe==0.7.0
networkx==3.4.2
openai==1.66.3
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
pylint==3.3.5
pyparsing==3.2.1
python-dotenv==1.0.1
sniffio==1.3.1
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
